Week 1 DONE: + Removed all warnings + Shortest path is now refreshed when cow and rabbit are have moved to a new node + Additional code documentation has been added + Shortest path is now completely shown on the SDL screen instead of the console window

+++ SDLFramework/SDLFramework/AStar.cpp
@@ -11,10 +11,13 @@ AStar::~AStar()
 {
 }
 
+// Method responsible for calculating the shortest path between a given start node and goal node
+// The method is based on the AStar algorithm pseudo code found on Wikipedia:
+// https://en.wikipedia.org/wiki/A*_search_algorithm
 stack<Node*> AStar::GetShortestPath(Node* start, Node* goal)
 {
-	this->source = start;
-	this->target = goal;
+	this->source = start;																		// Set the start node
+	this->target = goal;																		// Set the goal node
 
 	openList.insert(start);																		// The set of tentative nodes to be evaluated.
 	cameFrom[start->id] = nullptr;																// The map of navigated nodes.
@@ -26,21 +29,21 @@ stack<Node*> AStar::GetShortestPath(Node* start, Node* goal)
 	{
 		Node* current = (*openList.begin());													// the node in openset having the lowest f_score[] value
 
-		if (current == goal)																	 
-			return ReconstructPath(goal);
-
+		if (current == goal)																	// If the current node is equal to target goal,
+			return ReconstructPath(goal);														// Reconstruct and return the shortest path in a stack
 
 		openList.erase(current);																// remove from current openset
 		closedList.push_back(current);															// add  current to closedset
 
-		for (int i = 0; i < current->GetEdges().size(); i++)									// Check all the node's edges
+		for (size_t i = 0; i < current->GetEdges().size(); i++)									// Check all the node's edges
 		{
 			Node* next = current->GetEdges()[i]->child;											// The first edge's child node should be next node to be checked
 
 			if (find(closedList.begin(), closedList.end(), next) != closedList.end())			// If the 'next' node is found, then continue
 				continue;
 
-			double g_Cost = current->g_distance_to_source + current->GetEdges()[i]->weight;		// Calculate the G cost for the current node by
+			float g_Cost = static_cast<float>(current->g_distance_to_source + 
+				current->GetEdges()[i]->weight);												// Calculate the G cost for the current node by
 																								// adding the distance to source along with the weight of the edge
 
 			if (openList.find(next) == openList.end() || g_Cost < next->g_distance_to_source)	// If the next node is the same as the last one in the open list OR
@@ -51,15 +54,15 @@ stack<Node*> AStar::GetShortestPath(Node* start, Node* goal)
 					openList.erase(next);														
 
 				next->g_distance_to_source = g_Cost;											// Set the G Cost for the next node,
-				next->f_totalDistance = next->g_distance_to_source + CalculateH(next, goal);  // Calculate and set the F value by adding the G value and the H value
+				next->f_totalDistance = next->g_distance_to_source + CalculateH(next, goal);    // Calculate and set the F value by adding the G value and the H value
 																								// using the next node and the target node.
 
 				openList.insert(next);															// insert the next node in the open list
 			}
 		}
 	}
-	stack<Node*> failure;
-	return failure;
+	stack<Node*> failure;																		// If method has been unsuccessful at producing a shortest path
+	return failure;																				// Return an empty stack
 }
 
 float AStar::CalculateH(Node* start, Node* goal)
@@ -88,5 +91,5 @@ bool AStar::CompareNode::operator()(const Node* node1, const Node* node2) const
 	if (node1 == nullptr) return true;
 	if (node2 == nullptr) return false;
 
-	return *node1 < *node2; // Holly shit! DO NOT FORGET THOSE POINTER ASTERICKS! OTHERWISE THE WHOLE THING WILL NOT COMPARE PROPERLY!
+	return *node1 < *node2; // Wow ! DO NOT FORGET THOSE POINTER ASTERICKS! OTHERWISE THE WHOLE THING WILL NOT COMPARE PROPERLY!
 }
